Crate node_subtensor_runtime

Re-exports

pub use pallet_scheduler;

pub use pallet_subtensor;

Modules

api

check_nonce

opaque

Macros

construct_runtime

Construct a runtime, with the given name and the given pallets.

parameter_types

Create new implementations of the Get trait.

Structs

AbsoluteMinimumContribution

AllowCommitments

AllowIdentityReg

AnnouncementDepositBase

AnnouncementDepositFactor

AuraPalletIntrf

BaseFeeThreshold

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockGasLimit

BlockHashCount

BlockLength

BlockWeights

BoundDivision

CanProposeToTriumvirate

CanVoteToTriumvirate

CommitmentFieldDeposit

CommitmentInitialDeposit

ConfigurableChainId

ConstBool

Const getter for a basic type.

ConstU8

Const getter for a basic type.

ConstU32

Const getter for a basic type.

ConstU64

Const getter for a basic type.

ConstU128

Const getter for a basic type.

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

CrowdloanPalletId

DefaultBaseFeePerGas

DefaultElasticity

DepositBase

DepositFactor

DisallowPermissionlessEnterDuration

DisallowPermissionlessEntering

DisallowPermissionlessExtendDuration

DisallowPermissionlessExtending

DisallowPermissionlessRelease

DurationOfStartCall

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

FeeMultiplier

FieldDeposit

FindAuthorTruncated

GasLimitPovSizeRatio

GetSenateMemberCount

GrandpaInterfaceImpl

IdentityFee

Implementor of WeightToFee that maps one unit of weight to one unit of fee.

InitialAlphaHigh

InitialAlphaLow

InitialColdkeySwapRescheduleDuration

InitialColdkeySwapScheduleDuration

InitialDeposit

InitialDissolveNetworkScheduleDuration

InitialEmaPriceHalvingPeriod

InitialLiquidAlphaOn

InitialYuma3On

LinearWeightToFee

ManageSenateMembers

MaxAdditionalFields

MaxCommitFields

MaxCommitFieldsInner

MaxContributors

MaxPending

MaxProxies

MaxScheduledPerBlock

MaxSignatories

MaximumBlockDuration

MaximumSchedulerWeight

MinimumBlockDuration

MinimumDeposit

NoNestingCallFilter

NoPreimagePostponement

OperationalFeeMultiplier

OriginPrivilegeCmp

Used the compare the privilege of an origin inside the scheduler.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PostBlockAndTxnHashes

PrecompilesValue

PreimageBaseDeposit

PreimageByteDeposit

PreimageHoldReason

PreimageMaxSize

ProxyDepositBase

ProxyDepositFactor

RefundContributorsLimit

ResetBondsOnCommit

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

RootEnterDuration

RootExtendDuration

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SafeModeWhitelistedCalls

SenateMaxMembers

StorageInfo

Metadata about storage from the runtime.

SubtensorEvmBalanceConverter

SubtensorInitialActivityCutoff

SubtensorInitialAdjustmentAlpha

SubtensorInitialAdjustmentInterval

SubtensorInitialAlphaSigmoidSteepness

SubtensorInitialBondsMovingAverage

SubtensorInitialBondsPenalty

SubtensorInitialBondsResetOn

SubtensorInitialBurn

SubtensorInitialDefaultChildKeyTake

SubtensorInitialDefaultTake

SubtensorInitialDifficulty

SubtensorInitialEmissionValue

SubtensorInitialImmunityPeriod

SubtensorInitialIssuance

SubtensorInitialKappa

SubtensorInitialKeySwapCost

SubtensorInitialMaxAllowedUids

SubtensorInitialMaxAllowedValidators

SubtensorInitialMaxBurn

SubtensorInitialMaxChildKeyTake

SubtensorInitialMaxDifficulty

SubtensorInitialMaxRegistrationsPerBlock

SubtensorInitialMaxWeightsLimit

SubtensorInitialMinAllowedUids

SubtensorInitialMinAllowedWeights

SubtensorInitialMinBurn

SubtensorInitialMinChildKeyTake

SubtensorInitialMinDelegateTake

SubtensorInitialMinDifficulty

SubtensorInitialMinLockCost

SubtensorInitialNetworkImmunity

SubtensorInitialNetworkLockReductionInterval

SubtensorInitialNetworkRateLimit

SubtensorInitialPruningScore

SubtensorInitialRAORecycledForRegistration

SubtensorInitialRho

SubtensorInitialScalingLawPower

SubtensorInitialSenateRequiredStakePercentage

SubtensorInitialServingRateLimit

SubtensorInitialSubnetOwnerCut

SubtensorInitialTaoWeight

SubtensorInitialTargetRegistrationsPerInterval

SubtensorInitialTempo

SubtensorInitialTxChildKeyTakeRateLimit

SubtensorInitialTxDelegateTakeRateLimit

SubtensorInitialTxRateLimit

SubtensorInitialValidatorPruneLen

SubtensorInitialWeightsVersionKey

SuicideQuickClearLimit

TempoInterface

TransactionConverter

TransactionFeeHandler

Deduct the transaction fee from the Subtensor Pallet TotalIssuance when dropping the transaction fee.

TriumvirateVotes

Version

Weight

WeightPerGas

WeightToFeeCoefficient

One coefficient and its position in the WeightToFee.

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

EXISTENTIAL_DEPOSIT

INITIAL_CHILDKEY_TAKE_RATELIMIT

INITIAL_SUBNET_TEMPO

MAXIMUM_BLOCK_WEIGHT

RUNTIME_API_VERSIONS

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_REF_TIME_PER_SECOND

Traits

BuildStorage

Complex storage builder stuff.

FindAuthor

A trait for finding the author of a block header based on the PreRuntime digests contained within it.

InstanceFilter

Simple trait for providing a filter over a reference to some type, given an instance of itself.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

OnFinalize

See Hooks::on_finalize.

OnTimestampSet

A trait which is called when the timestamp is set in the runtime.

PrivilegeCmp

Something that can compare privileges of two origins.

Randomness

A trait that is able to provide randomness.

StorageValue

A trait for working with macro-generated storage values under the substrate storage API.

WeightToFeePolynomial

A trait that describes the weight to fee calculation as polynomial.

Functions

deposit

native_version

Type Aliases

Address

AdminUtils

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

Balances

BalancesConfig

BaseFee

BaseFeeConfig

Block

CheckedExtrinsic

Extrinsic type that has already been checked.

Commitments

Crowdloan

Drand

EVM

EVMChainId

EVMChainIdConfig

EVMConfig

Ethereum

EthereumConfig

Executive

Grandpa

GrandpaConfig

Header

Multisig

Preimage

Proxy

RandomnessCollectiveFlip

Registry

SafeMode

SafeModeConfig

Scheduler

SenateMembers

SenateMembersConfig

SignedExtra

SignedPayload

SubtensorModule

SubtensorModuleConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Triumvirate

TriumvirateConfig

TriumvirateMembers

TriumvirateMembersConfig

UncheckedExtrinsic

Utility

WeightToFeeCoefficients

A list of coefficients that represent a polynomial.