node_subtensor_runtime

Trait FindAuthor

pub trait FindAuthor<Author> {
    // Required method
    fn find_author<'a, I>(digests: I) -> Option<Author>
       where I: 'a + IntoIterator<Item = ([u8; 4], &'a [u8])>;
}

A trait for finding the author of a block header based on the PreRuntime digests contained within it.

Required Methods

fn find_author<'a, I>(digests: I) -> Option<Author>
where I: 'a + IntoIterator<Item = ([u8; 4], &'a [u8])>,

Find the author of a block based on the pre-runtime digests.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.