node_subtensor_runtime

Trait PrivilegeCmp

pub trait PrivilegeCmp<Origin> {
    // Required method
    fn cmp_privilege(left: &Origin, right: &Origin) -> Option<Ordering>;
}

Something that can compare privileges of two origins.

Required Methods

fn cmp_privilege(left: &Origin, right: &Origin) -> Option<Ordering>

Compare the left to the right origin.

The returned ordering should be from the pov of the left origin.

Should return None when it can not compare the given origins.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.