node_subtensor_runtime

Type Alias AllPalletsWithSystem

pub type AllPalletsWithSystem = (System, RandomnessCollectiveFlip, Timestamp, Aura, Grandpa, Balances, TransactionPayment, SubtensorModule, Triumvirate, TriumvirateMembers, SenateMembers, Utility, Sudo, Multisig, Preimage, Scheduler, Proxy, Registry, Commitments, AdminUtils, SafeMode, Ethereum, EVM, EVMChainId, BaseFee, Drand, Crowdloan);

All pallets included in the runtime as a nested tuple of types.