node_subtensor_runtime

Type Alias CheckedExtrinsic

pub type CheckedExtrinsic = CheckedExtrinsic<AccountId, RuntimeCall, SignedExtra, H160>;

Extrinsic type that has already been checked.

Aliased Type

struct CheckedExtrinsic {
    pub signed: CheckedSignature<AccountId32, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>, SubtensorSignedExtension<Runtime>, CommitmentsSignedExtension<Runtime>, CheckMetadataHash<Runtime>), H160>,
    pub function: RuntimeCall,
}

Fields
signed: CheckedSignature<AccountId32, (CheckNonZeroSender<Runtime>, CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>, SubtensorSignedExtension<Runtime>, CommitmentsSignedExtension<Runtime>, CheckMetadataHash<Runtime>), H160>

Who this purports to be from and the number of extrinsics have come before from the same signer, if anyone (note this is not a signature).

function: RuntimeCall

The function that should be called.